Added an available lab tests tab to LT
The "Available Lab Tests" tab is to help users find the ID # of all of the available lab tests that the clinic offers, by displaying all of the lab tests and their IDs that a physician can request. This can be helpful when the user is requesting a new lab test. (similar to the patient list on the ER)